% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_findIsotopes.R
\name{calcIsotopeMatrix}
\alias{calcIsotopeMatrix}
\title{Calculate an isotope matrix

\code{calcIsotopeMatrix} creates a matrix which is used to identify isotopes.}
\usage{
calcIsotopeMatrix(maxiso = 4L)
}
\arguments{
\item{maxiso}{\code{integer(1)} defining the maximum number of isotopes that
should be searched for.}
}
\value{
\code{matrix} with 4 columns \code{"mzmin"}, \code{"mzmax"}, \code{"intmin"} and
\code{"intmax"} defining the lower and upper boundaries for m/z and intensity
differences expected for isotopes
}
\description{
Calculate an isotope matrix

\code{calcIsotopeMatrix} creates a matrix which is used to identify isotopes.
}
