% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00CAMERA.R
\docType{class}
\name{CAMERA}
\alias{CAMERA}
\title{R6 class for CAMERA}
\description{
A simple wrapper function.
Using a summary set, identify set of instruments for the traits, and perform SEM MR to test the association across the population.
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{output}}{A list for the output}

    \item{\code{source}}{The source of the data.}

    \item{\code{exposure_ids}}{Exposures IDs obtained from IEU GWAS database (\url{https://gwas.mrcieu.ac.uk}) for each population}

    \item{\code{outcome_ids}}{Outcome IDs obtained from IEU GWAS database (\url{https://gwas.mrcieu.ac.uk}) for each population}

    \item{\code{exposure_metadata}}{Exposure metadata}

    \item{\code{outcome_metadata}}{Outcome metadata}

    \item{\code{radius}}{Genomic window size to extract SNPs}

    \item{\code{pops}}{Ancestry information for each population (i.e., AFR, AMR, EUR, EAS, SAS)}

    \item{\code{bfiles}}{Locations of LD reference files for each population (Download from: \url{http://fileserve.mrcieu.ac.uk/ld/1kg.v3.tgz})}

    \item{\code{plink}}{Location of executable plink (version 1.90 is recommended)}

    \item{\code{clump_pop}}{Reference population for clumping}

    \item{\code{instrument_raw}}{Instruments for the exposures obtained from \code{extract_instruments()}}

    \item{\code{instrument_regions}}{Genomic regions around each instrument obtained from \code{extract_instrument_regions()}}

    \item{\code{ld_matrices}}{LD matrices for each instrument region obtained from \code{regional_ld_matrices()}}

    \item{\code{susie_results}}{Fine-mapping results from \code{susie_finemap_regions()}}

    \item{\code{paintor_results}}{Fine-mapping results from \code{paintor_finemap_regions()}}

    \item{\code{mscaviar_results}}{Fine-mapping results from \code{MsCAVIAR_finemap_regions()}}

    \item{\code{expected_replications}}{Expected replication of instruments across populations}

    \item{\code{instrument_region_zscores}}{Z scores for the SNPs in each instrument region obtained from \code{scan_regional_instruments()}}

    \item{\code{instrument_fema}}{Instruments selected by meta-analysis across populations obtained from \code{fema_regional_instruments()}}

    \item{\code{instrument_fema_regions}}{Meta-analysis results for each instrument region obtained from \code{fema_regional_instruments()}}

    \item{\code{instrument_maxz}}{Instruments selected by maximum Z score obtained from \code{scan_regional_instruments()}}

    \item{\code{instrument_susie}}{Instruments selected by susieR fine-mapping obtained from \code{susie_finemap_regions()}}

    \item{\code{instrument_paintor}}{Instruments selected by PAINTOR fine-mapping obtained from \code{paintor_finemap_regions()}}

    \item{\code{instrument_mscaviar}}{Instruments selected by MsCAVIAR fine-mapping obtained from \code{MsCAVIAR_finemap_regions()}}

    \item{\code{harmonised_data_check}}{Checks of the harmonised data}

    \item{\code{standardised_instrument_raw}}{Standardised \code{instrument_raw} obtained from \code{standardise_data()}}

    \item{\code{standardised_instrument_maxz}}{Standardised \code{instrument_maxz} obtained from \code{standardise_data()}}

    \item{\code{standardised_instrument_susie}}{Standardised \code{instrument_susie} obtained from \code{standardise_data()}}

    \item{\code{standardised_instrument_paintor}}{Standardised \code{instrument_paintor} obtained from \code{standardise_data()}}

    \item{\code{standardised_instrument_mscaviar}}{Standardised \code{instrument_mscaviar} obtained from \code{standardise_data()}}

    \item{\code{standardised_outcome}}{Standardised outcome data obtained from \code{standardise_data()}}

    \item{\code{instrument_specificity}}{Instrument specificity results obtained from \code{estimate_instrument_specificity()}}

    \item{\code{instrument_specificity_summary}}{Summary of the instrument specificity results obtained from \code{estimate_instrument_specificity()}}

    \item{\code{instrument_outcome}}{Outcome data for the instruments obtained from \code{make_outcome_data()}}

    \item{\code{instrument_outcome_regions}}{Outcome data for each instrument region imported with \code{import_from_local()}}

    \item{\code{harmonised_dat_sem}}{Harmonised data for the SEM analysis}

    \item{\code{harmonised_dat}}{Harmonised exposure and outcome data obtained from \code{harmonise()}}

    \item{\code{sem_result}}{SEM results obtained from \code{perform_basic_sem()}}

    \item{\code{pleiotropy_agreement}}{Agreement of pleiotropy outlier effects across populations obtained from \code{pleiotropy()}}

    \item{\code{pleiotropy_outliers}}{Pleiotropy outliers obtained from \code{pleiotropy()}}

    \item{\code{pleiotropy_Q_outliers}}{Heterogeneity for each pleiotropy outlier and population obtained from \code{pleiotropy()}}

    \item{\code{summary}}{Summary of the exposure, outcome and population metadata obtained from \code{set_summary()}}

    \item{\code{mrres}}{MR estimates obtained from \code{cross_estimate()}}

    \item{\code{instrument_heterogeneity_per_variant}}{Heterogeneity of the instrument-exposure associations for each variant obtained from \code{estimate_instrument_heterogeneity_per_variant()}}

    \item{\code{mrgxe_res}}{MR GxE results obtained from \code{mrgxe()}}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-CAMERA-import}{\code{CAMERA$import()}}
    \item \href{#method-CAMERA-assign}{\code{CAMERA$assign()}}
    \item \href{#method-CAMERA-import_from_local}{\code{CAMERA$import_from_local()}}
    \item \href{#method-CAMERA-initialize}{\code{CAMERA$new()}}
    \item \href{#method-CAMERA-instrument_heterogeneity}{\code{CAMERA$instrument_heterogeneity()}}
    \item \href{#method-CAMERA-estimate_instrument_specificity}{\code{CAMERA$estimate_instrument_specificity()}}
    \item \href{#method-CAMERA-replication_evaluation}{\code{CAMERA$replication_evaluation()}}
    \item \href{#method-CAMERA-check_phenotypes}{\code{CAMERA$check_phenotypes()}}
    \item \href{#method-CAMERA-cross_estimate}{\code{CAMERA$cross_estimate()}}
    \item \href{#method-CAMERA-plot_cross_estimate}{\code{CAMERA$plot_cross_estimate()}}
    \item \href{#method-CAMERA-extract_instruments}{\code{CAMERA$extract_instruments()}}
    \item \href{#method-CAMERA-extract_instrument_regions}{\code{CAMERA$extract_instrument_regions()}}
    \item \href{#method-CAMERA-scan_regional_instruments}{\code{CAMERA$scan_regional_instruments()}}
    \item \href{#method-CAMERA-plot_regional_instruments_maxz}{\code{CAMERA$plot_regional_instruments_maxz()}}
    \item \href{#method-CAMERA-regional_ld_matrices}{\code{CAMERA$regional_ld_matrices()}}
    \item \href{#method-CAMERA-susie_finemap_regions}{\code{CAMERA$susie_finemap_regions()}}
    \item \href{#method-CAMERA-paintor_finemap_regions}{\code{CAMERA$paintor_finemap_regions()}}
    \item \href{#method-CAMERA-MsCAVIAR_finemap_regions}{\code{CAMERA$MsCAVIAR_finemap_regions()}}
    \item \href{#method-CAMERA-fema_regional_instruments}{\code{CAMERA$fema_regional_instruments()}}
    \item \href{#method-CAMERA-plot_regional_instruments}{\code{CAMERA$plot_regional_instruments()}}
    \item \href{#method-CAMERA-get_metadata}{\code{CAMERA$get_metadata()}}
    \item \href{#method-CAMERA-estimate_instrument_heterogeneity_per_variant}{\code{CAMERA$estimate_instrument_heterogeneity_per_variant()}}
    \item \href{#method-CAMERA-mrgxe}{\code{CAMERA$mrgxe()}}
    \item \href{#method-CAMERA-mrgxe_plot}{\code{CAMERA$mrgxe_plot()}}
    \item \href{#method-CAMERA-mrgxe_plot_variant}{\code{CAMERA$mrgxe_plot_variant()}}
    \item \href{#method-CAMERA-make_outcome_data}{\code{CAMERA$make_outcome_data()}}
    \item \href{#method-CAMERA-make_outcome_local}{\code{CAMERA$make_outcome_local()}}
    \item \href{#method-CAMERA-harmonise}{\code{CAMERA$harmonise()}}
    \item \href{#method-CAMERA-set_summary}{\code{CAMERA$set_summary()}}
    \item \href{#method-CAMERA-pleiotropy}{\code{CAMERA$pleiotropy()}}
    \item \href{#method-CAMERA-plot_pleiotropy}{\code{CAMERA$plot_pleiotropy()}}
    \item \href{#method-CAMERA-plot_pleiotropy_heterogeneity}{\code{CAMERA$plot_pleiotropy_heterogeneity()}}
    \item \href{#method-CAMERA-perform_basic_sem}{\code{CAMERA$perform_basic_sem()}}
    \item \href{#method-CAMERA-runsem}{\code{CAMERA$runsem()}}
    \item \href{#method-CAMERA-standardise_data}{\code{CAMERA$standardise_data()}}
    \item \href{#method-CAMERA-clone}{\code{CAMERA$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-import"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-import}{}}}
\subsection{\code{CAMERA$import()}}{
  Migrate the results from a previous CAMERA
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$import(x)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{x}}{A CAMERA object to import results from}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-assign"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-assign}{}}}
\subsection{\code{CAMERA$assign()}}{
  Assign values to fields of the CAMERA object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$assign(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Named arguments, where each name is a field and each value is assigned to it}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-import_from_local"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-import_from_local}{}}}
\subsection{\code{CAMERA$import_from_local()}}{
  Import instrument and outcome data from local summary statistics, e.g. as organised by \code{CAMERA_local}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$import_from_local(
  instrument_raw,
  instrument_outcome,
  instrument_regions,
  instrument_outcome_regions,
  exposure_ids,
  outcome_ids,
  pops,
  ...
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{instrument_raw}}{Data frame of instruments for the exposures}
      \item{\code{instrument_outcome}}{Data frame of instrument associations with the outcomes}
      \item{\code{instrument_regions}}{List of genomic regions around each instrument for the exposures}
      \item{\code{instrument_outcome_regions}}{List of genomic regions around each instrument for the outcomes}
      \item{\code{exposure_ids}}{Exposures IDs obtained from IEU GWAS database (\url{https://gwas.mrcieu.ac.uk}) for each population}
      \item{\code{exposure_ids}}{ID for the exposure. Default is x$exposure_ids.}
      \item{\code{exposure_ids}}{IDs for the exposures. Default is \code{x$exposure_ids}.}
      \item{\code{outcome_ids}}{Outcome IDs obtained from IEU GWAS database (\url{https://gwas.mrcieu.ac.uk}) for each population}
      \item{\code{outcome_ids}}{IDs for the outcomes. Default is \code{x$outcome_ids}.}
      \item{\code{pops}}{Ancestry information for each population (i.e., AFR, AMR, EUR, EAS, SAS)}
      \item{\code{...}}{Further named arguments passed to \code{assign()}}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-initialize}{}}}
\subsection{\code{CAMERA$new()}}{
  Create a new dataset and initialise an R interface
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$new(
  exposure_ids = NULL,
  outcome_ids = NULL,
  pops = NULL,
  bfiles = NULL,
  plink = NULL,
  radius = NULL,
  clump_pop = NULL,
  x = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{exposure_ids}}{Exposures IDs obtained from IEU GWAS database (\url{https://gwas.mrcieu.ac.uk}) for each population}
      \item{\code{exposure_ids}}{ID for the exposure. Default is x$exposure_ids.}
      \item{\code{exposure_ids}}{IDs for the exposures. Default is \code{x$exposure_ids}.}
      \item{\code{outcome_ids}}{Outcome IDs obtained from IEU GWAS database (\url{https://gwas.mrcieu.ac.uk}) for each population}
      \item{\code{outcome_ids}}{IDs for the outcomes. Default is \code{x$outcome_ids}.}
      \item{\code{pops}}{Ancestry information for each population (i.e., AFR, AMR, EUR, EAS, SAS)}
      \item{\code{bfiles}}{Locations of LD reference files for each population (Download from: \url{http://fileserve.mrcieu.ac.uk/ld/1kg.v3.tgz})}
      \item{\code{plink}}{Location of executable plink (version 1.90 is recommended)}
      \item{\code{radius}}{Genomic window size to extract SNPs}
      \item{\code{clump_pop}}{Reference population for clumping}
      \item{\code{x}}{Import data where available}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-instrument_heterogeneity"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-instrument_heterogeneity}{}}}
\subsection{\code{CAMERA$instrument_heterogeneity()}}{
  The function evaluates heterogeneity in the association of selected instruments and the exposure/outcome between the populations. Heterogeneity (Q statistics) is calcuated based on an IVW or simple MODE MR estimator. The instruments can be identified using "Raw", "MaxZ", or fine-mapping (Susie, PAINTOR) methods.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$instrument_heterogeneity(
  instrument = self$instrument_raw,
  alpha = "bonferroni",
  method = "ivw",
  outlier_removal = FALSE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{instrument}}{Intsruments for the exposure that are selected by using the provided methods in CAMERA (x$instrument_raw, x$instrument_maxz, x$instrument_susie, x$instrument_paintor). Default is x$instrument_raw.}
      \item{\code{alpha}}{Statistical threshold to determine significance. Default is "bonferroni", which is eqaul to 0.05/number of the instruments.
For each pair of populations, the instruments that are significant (\code{p < alpha}) in the reference population are taken and their associations in the replication population are regressed on their associations in the reference population. If the instruments have the same effects in both populations the slope (\code{agreement}) is expected to be close to 1.}
      \item{\code{method}}{Method used to estimate the agreement slope, either \code{"ivw"} (inverse variance weighted, using \code{\link[TwoSampleMR:mr_ivw]{TwoSampleMR::mr_ivw()}}) or \code{"simple_mode"} (using \code{\link[TwoSampleMR:mr_simple_mode]{TwoSampleMR::mr_simple_mode()}}). Default is \code{"ivw"}.}
      \item{\code{outlier_removal}}{Remove outliers identified by radial IVW MR before estimating heterogeneity. Only the first two exposures are used to identify outliers. Default is \code{FALSE}.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Data frame with one row for each ordered pair of populations, with columns:
\itemize{
\item \code{Reference}, \code{Replication}: IDs of the reference and replication datasets
\item \code{nsnp}: number of instruments significant in the reference dataset that are also present in the replication dataset
\item \code{agreement}: slope of the regression of the replication associations on the reference associations
\item \code{se}, \code{pval}: standard error of \code{agreement} and p-value for the test of the null hypothesis that the slope is 0 (i.e. no association between the instrument associations in the two populations)
\item \code{Q}, \code{Q_pval} (\code{"ivw"} only): Cochran's Q statistic and its p-value for the null hypothesis that all instruments share the same slope, i.e. no heterogeneity in the agreement between populations
\item \code{I2} (\code{"ivw"} only): the proportion of the variation in the ratio of the associations that is due to heterogeneity rather than chance, calculated as \code{(Q - nsnp) / Q}
}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-estimate_instrument_specificity"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-estimate_instrument_specificity}{}}}
\subsection{\code{CAMERA$estimate_instrument_specificity()}}{
  The function estimates what fraction of the instuments is expected to be replicated across the populations under the hypothesis that the effect estimates are the same.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$estimate_instrument_specificity(
  instrument,
  alpha = "bonferroni",
  winnerscurse = FALSE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{instrument}}{Intsruments for the exposure that are selected by using the provided methods in CAMERA (x$instrument_raw, x$instrument_maxz, x$instrument_susie, x$instrument_paintor). Default is x$instrument_raw.}
      \item{\code{alpha}}{Statistical threshold to determine significance. Default is "bonferroni", which is eqaul to 0.05/number of the instruments.}
      \item{\code{winnerscurse}}{Use this option to correct winners' curse bias.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Details}{
    For each pair of populations, the instruments that are significant (\code{p < alpha}) in the discovery population are compared with their associations in the replication population. Following Okbay et al. (2016), under the hypothesis that the instruments have the same effects in both populations, the expected number of instruments with the same sign, and the expected number with \code{p < alpha}, in the replication population are calculated taking account of the precision of the replication associations. These are compared with the observed numbers.
  }
  \subsection{Returns}{
    Data frame, also stored in \code{x$instrument_specificity_summary}, with columns:
\itemize{
\item \code{discovery}, \code{replication}: IDs of the discovery and replication datasets
\item \code{metric}: \code{"Sign"} (the instrument has the same sign in both populations) or \code{"P-value"} (the instrument has \code{p < alpha} in the replication population)
\item \code{nsnp}: number of instruments compared
\item \code{datum}, \code{value}: the \code{"Expected"} and \code{"Observed"} number of instruments meeting \code{metric}
\item \code{pdiff}: p-value from a binomial test of the null hypothesis that the observed number equals the expected number. A small p-value suggests the instruments replicate less (or more) than expected if their effects were the same in both populations.
}

Per variant results are stored in \code{x$instrument_specificity}, including the expected probabilities of sign agreement (\code{sign}) and replication (\code{sig}), and \code{distinct}, which flags instruments that were expected to replicate but did not (\code{sig > 0.8} and replication \code{p > 0.1}), or were expected to have the same sign but did not (\code{sign > 0.8}).
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-replication_evaluation"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-replication_evaluation}{}}}
\subsection{\code{CAMERA$replication_evaluation()}}{
  The function explains what contributes to the replication of gene-trait association between the populations, considering LD structure.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$replication_evaluation(
  instrument = self$instrument_raw,
  ld = self$ld_matrices
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{instrument}}{Intsruments for the exposure that are selected by using the provided methods in CAMERA (x$instrument_raw, x$instrument_maxz, x$instrument_susie, x$instrument_paintor). Default is x$instrument_raw.}
      \item{\code{ld}}{LD matrix obtained by using \code{x$regional_ld_matrices()}}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Table of the regression result
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-check_phenotypes"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-check_phenotypes}{}}}
\subsection{\code{CAMERA$check_phenotypes()}}{
  This function evaluates how competitable genetic associations in population 1 are with those in population 2. The function checks whether 1) the chosen IDs for the exposure or outcome can be used for the further steps and 2) the units for the genetic associations are comparable between populations.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$check_phenotypes(ids = self$exposure_ids)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{ids}}{ID for the exposure or the outcome. Default is x$exposure_ids.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Table of the result.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-cross_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-cross_estimate}{}}}
\subsection{\code{CAMERA$cross_estimate()}}{
  This method performs a IVW analysis on the harmonised data. It fits two linear models, one considering all populations and another considering each population separately. It then performs a fixed effects meta-analysis to estimate the heterogeneity across different populations. The results are stored in the \code{mrres} attribute of the \code{CAMERA} object.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$cross_estimate(dat = self$harmonised_dat)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dat}}{A data frame containing the harmonised data. It should have the columns \code{beta.y}, \code{beta.x}, \code{se.y}, and \code{pops}. If not provided, the method uses the \code{harmonised_dat} attribute of the \code{CAMERA} object.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Data frame, also stored in \code{x$mrres}, with one row for the estimate across all populations (\code{pops == "All"}) and one row for each population, with columns:
\itemize{
\item \code{pops}: population
\item \code{Estimate}, \verb{Std. Error}, \verb{t value}, \verb{Pr(>|t|)}: the IVW MR estimate, i.e. the weighted regression of the SNP-outcome associations on the SNP-exposure associations through the origin with weights \code{1 / se.y^2}, its standard error, t statistic and p-value
\item \code{Qj}, \code{Qjpval}, \code{Qdf}: for each population, the contribution of its estimate to the heterogeneity between populations and its p-value (1 degree of freedom), from a fixed effects meta analysis of the population estimates; a small p-value suggests the estimate in that population differs from the others. For the \code{"All"} row these are Cochran's Q statistic across all populations, its p-value (null hypothesis: the same causal effect in all populations) and its degrees of freedom.
}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-plot_cross_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-plot_cross_estimate}{}}}
\subsection{\code{CAMERA$plot_cross_estimate()}}{
  Plot the results from \code{cross_estimate}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$plot_cross_estimate(est = self$mrres, qj_alpha = 0.05)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{est}}{Results from \code{cross_estimate()}. Default is \code{x$mrres}.}
      \item{\code{qj_alpha}}{Significance threshold for highlighting heterogeneous estimates. Default is 0.05.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Plot
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-extract_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-extract_instruments}{}}}
\subsection{\code{CAMERA$extract_instruments()}}{
  This function searches for GWAS significant SNPs (P < 5E-8) for a specified set of the exposures. This method is equivalant to the instrumnet extraction method for Multivariable MR. Reference here: https://mrcieu.github.io/TwoSampleMR/reference/mv_extract_exposures.html.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$extract_instruments(
  exposure_ids = self$exposure_ids,
  harmonise_strictness = 1,
  ...
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{exposure_ids}}{ID for the exposure. Default is x$exposure_ids.}
      \item{\code{harmonise_strictness}}{Passed to \code{TwoSampleMR::mv_extract_exposures()}. Default is 1, which assumes all alleles are on the forward strand, as they are in OpenGWAS. The TwoSampleMR default of 2 infers the strand of palindromic SNPs from their allele frequencies, which wrongly flips palindromic SNPs whose allele frequencies are on different sides of 0.5 in different populations, and drops palindromic SNPs with intermediate allele frequencies.}
      \item{\code{...}}{Further arguments passed to \code{TwoSampleMR::mv_extract_exposures()}}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Data frame in x$instrument_raw
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-extract_instrument_regions"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-extract_instrument_regions}{}}}
\subsection{\code{CAMERA$extract_instrument_regions()}}{
  This function extract genomic regions around each instrument (e.g. 50kb) obtained from \code{x$extract_instruments()}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$extract_instrument_regions(
  radius = self$radius,
  instrument_raw = self$instrument_raw,
  exposure_ids = self$exposure_ids
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{radius}}{Set a range of the region to search}
      \item{\code{instrument_raw}}{A set of instruments obtained from \code{x$extract_instruments()}}
      \item{\code{exposure_ids}}{ID for the exposure. Default is x$exposure_ids.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Data frame in x$instrument_regions
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-scan_regional_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-scan_regional_instruments}{}}}
\subsection{\code{CAMERA$scan_regional_instruments()}}{
  The function searches for a SNP that is best associated across the poupulations within the identified regions by using \code{x$extract_instrument_regions()}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$scan_regional_instruments(
  instrument_raw = self$instrument_raw,
  instrument_regions = self$instrument_regions
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{instrument_raw}}{Instruments for the exposures by using \code{x$extract_instrument()}}
      \item{\code{instrument_regions}}{Genomic regions identified by using \code{x$extract_instrument_regions()}}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    List of z scores for the chosen SNPs in x$instrument_region_zscores. Data frame in x$instrument_maxz
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-plot_regional_instruments_maxz"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-plot_regional_instruments_maxz}{}}}
\subsection{\code{CAMERA$plot_regional_instruments_maxz()}}{
  The function draws a plot shows how CAMERA selected the instruments for the exposure from the selected genomic regions.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$plot_regional_instruments_maxz(
  instrument_region_zscores = self$instrument_region_zscores,
  instruments = self$instrument_raw,
  region = 1:min(10, nrow(instruments)),
  comparison = FALSE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{instrument_region_zscores}}{Y axis. Z scores based on the SNP-exposure associations.}
      \item{\code{instruments}}{Use this option to draw a separate plot for the selcted instruments.}
      \item{\code{region}}{X axis. Number of genomic regions to be shown in the plot. Default is 10.}
      \item{\code{comparison}}{Use this option to compare the selected instruments by different instrument selection methods in one plot.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Plot
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-regional_ld_matrices"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-regional_ld_matrices}{}}}
\subsection{\code{CAMERA$regional_ld_matrices()}}{
  Generate LD matrices for instrument regions.
If we want to do fine mapping we need to get an LD matrix for the whole region (for each population)
We then need to harmonise the LD matrix to the summary data, and the summary datasets to each other
The function obtains an LD matrix for the selected genomic regions.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$regional_ld_matrices(
  instrument_regions = self$instrument_regions,
  bfiles = self$bfiles,
  pops = self$pops,
  plink = self$plink
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{instrument_regions}}{Genomic regions identified by using \code{x$extract_instrument_regions()}}
      \item{\code{bfiles}}{Location of LD reference files for each population (Download from: http://fileserve.mrcieu.ac.uk/ld/1kg.v3.tgz)}
      \item{\code{pops}}{Ancestry information for each population (i.e. AFR, AMR, EUR, EAS, SAS)}
      \item{\code{plink}}{Location of executable plink (version 1.90 is recommended)}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Data frame of LD matrix (x$ld_matrices)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-susie_finemap_regions"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-susie_finemap_regions}{}}}
\subsection{\code{CAMERA$susie_finemap_regions()}}{
  Fine-mapping using susieR (https://github.com/stephenslab/susieR) to extract instruments for the exposure for multiple populations. The function identifies credible sets that overlap between the populations and determine the best SNP in a genomic region to be used as instrument.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$susie_finemap_regions(
  dat = self$instrument_regions,
  ld = self$ld_matrices
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dat}}{Genomic regions identified by using \code{x$extract_instrument_regions()}}
      \item{\code{ld}}{LD matrix obtained by using \code{x$regional_ld_matrices()}}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Result from susieR in x$susie_results. Data frame in x$instrument_susie
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-paintor_finemap_regions"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-paintor_finemap_regions}{}}}
\subsection{\code{CAMERA$paintor_finemap_regions()}}{
  Fine-mapping using PAINTOR (https://github.com/gkichaev/PAINTOR_V3.0) to extract instruments for the exposure for multiple populations. The function chooses the SNP with highest posterior probability and associations in each exposure.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$paintor_finemap_regions(
  region = self$instrument_regions,
  ld = self$ld_matrices,
  PAINTOR = "PAINTOR",
  workdir = tempdir()
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{region}}{Genomic regions identified by using \code{x$extract_instrument_regions()}}
      \item{\code{ld}}{LD matrix obtained by using \code{x$regional_ld_matrices()}}
      \item{\code{PAINTOR}}{Path to executable PAINTOR. Default="PAINTOR"}
      \item{\code{workdir}}{Working directory to save the output. Default=tempdir()}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Result from PAINTOR in x$paintor_results. Data frame in x$instrument_paintor
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-MsCAVIAR_finemap_regions"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-MsCAVIAR_finemap_regions}{}}}
\subsection{\code{CAMERA$MsCAVIAR_finemap_regions()}}{
  Fine-mapping using MsCAVIAR (https://github.com/nlapier2/MsCAVIAR) to extract instruments for the exposure for multiple populations.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$MsCAVIAR_finemap_regions(
  region = self$instrument_regions,
  ld = self$ld_matrices,
  MsCAVIAR = "MsCAVIAR",
  workdir = tempdir()
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{region}}{Genomic regions identified by using \code{x$extract_instrument_regions()}}
      \item{\code{ld}}{LD matrix obtained by using \code{x$regional_ld_matrices()}}
      \item{\code{MsCAVIAR}}{Path to executable MsCAVIAR. Default="MsCAVIAR"}
      \item{\code{workdir}}{Working directory to save the output. Default=tempdir()}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Result from MsCAVIAR in x$mscaviar_results. Data frame in x$instrument_mscaviar
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-fema_regional_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-fema_regional_instruments}{}}}
\subsection{\code{CAMERA$fema_regional_instruments()}}{
  Identify the best variant for each region by meta-analysing the associations across the populations
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$fema_regional_instruments(
  method = "fema",
  instrument_regions = self$instrument_regions,
  instrument_raw = self$instrument_raw,
  n = self$exposure_metadata$sample_size
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{method}}{Meta-analysis method, either \code{"fema"} (fixed effects meta-analysis) or \code{"zma"} (Z score meta-analysis). Default is \code{"fema"}.}
      \item{\code{instrument_regions}}{Genomic regions identified by using \code{x$extract_instrument_regions()}}
      \item{\code{instrument_raw}}{Instruments for the exposures obtained from \code{x$extract_instruments()}}
      \item{\code{n}}{Sample sizes for each exposure, required for \code{method = "zma"}. Default is the sample sizes from \code{x$exposure_metadata}.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Data frame of the selected instruments, also stored in \code{x$instrument_fema}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-plot_regional_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-plot_regional_instruments}{}}}
\subsection{\code{CAMERA$plot_regional_instruments()}}{
  Plot the associations in a region for each population and for the meta-analysis
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$plot_regional_instruments(
  region,
  instrument_regions = self$instrument_regions,
  meta_analysis_regions = self$instrument_fema_regions
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{region}}{Name of the region to plot, i.e. one of \code{names(x$instrument_regions)}}
      \item{\code{instrument_regions}}{Genomic regions identified by using \code{x$extract_instrument_regions()}}
      \item{\code{meta_analysis_regions}}{Meta-analysis results for each region obtained from \code{x$fema_regional_instruments()}}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Plot
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-get_metadata}{}}}
\subsection{\code{CAMERA$get_metadata()}}{
  Get metadata for the exposures and outcomes from the OpenGWAS API
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$get_metadata(
  exposure_ids = self$exposure_ids,
  outcome_ids = self$outcome_ids
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{exposure_ids}}{IDs for the exposures. Default is \code{x$exposure_ids}.}
      \item{\code{outcome_ids}}{IDs for the outcomes. Default is \code{x$outcome_ids}.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    List of the exposure and outcome metadata, also stored in \code{x$exposure_metadata} and \code{x$outcome_metadata}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-estimate_instrument_heterogeneity_per_variant"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-estimate_instrument_heterogeneity_per_variant}{}}}
\subsection{\code{CAMERA$estimate_instrument_heterogeneity_per_variant()}}{
  Test for heterogeneity of effect estimates between populations.
For each SNP this function will provide a Cochran's Q test statistic - a measure of heterogeneity of effect sizes between populations. A low p-value means high heterogeneity.
In addition, for every SNP it gives a per population p-value - this can be interpreted as asking for each SNP is a particular giving an outlier estimate.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$estimate_instrument_heterogeneity_per_variant(dat = self$harmonised_dat)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dat}}{Harmonised data. Default is \code{x$harmonised_dat}.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Data frame of the Q statistic, its p-value and FDR adjusted p-value for each SNP, also stored in \code{x$instrument_heterogeneity_per_variant}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-mrgxe"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-mrgxe}{}}}
\subsection{\code{CAMERA$mrgxe()}}{
  Estimate the pleiotropic effect of each variant using MR GxE across populations, see \code{egger_bootstrap()}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$mrgxe(
  dat = self$harmonised_dat,
  variant_list = subset(self$instrument_heterogeneity_per_variant, Qfdr < 0.05)$SNP,
  nboot = 100
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dat}}{Harmonised data. Default is \code{x$harmonised_dat}.}
      \item{\code{variant_list}}{SNPs to analyse. Default is the SNPs with FDR < 0.05 in \code{x$instrument_heterogeneity_per_variant}.}
      \item{\code{nboot}}{Number of bootstraps. Default is 100.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Data frame of the MR GxE results, also stored in \code{x$mrgxe_res}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-mrgxe_plot"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-mrgxe_plot}{}}}
\subsection{\code{CAMERA$mrgxe_plot()}}{
  Plot the MR GxE pleiotropy estimates for each variant
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$mrgxe_plot(mrgxe_res = self$mrgxe_res)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{mrgxe_res}}{Results from \code{mrgxe()}. Default is \code{x$mrgxe_res}.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Plot
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-mrgxe_plot_variant"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-mrgxe_plot_variant}{}}}
\subsection{\code{CAMERA$mrgxe_plot_variant()}}{
  Plot the instrument-exposure against instrument-outcome associations across populations for selected variants
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$mrgxe_plot_variant(
  variant = self$mrgxe_res \%>\% dplyr::filter(p.adjust(a_pval, "fdr") < 0.05) \%>\% {
  
      .$SNP
 },
  dat = self$harmonised_dat
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{variant}}{SNPs to plot. Default is the SNPs with FDR < 0.05 for the pleiotropy estimate in \code{x$mrgxe_res}.}
      \item{\code{dat}}{Harmonised data. Default is \code{x$harmonised_dat}.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Plot
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-make_outcome_data"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-make_outcome_data}{}}}
\subsection{\code{CAMERA$make_outcome_data()}}{
  The function extracts summary statistics of given a list of instruments and the outcomes
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$make_outcome_data(exp = self$instrument_raw, p_exp = 0.05/nrow(exp))}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{exp}}{Intsruments for the exposure that are selected by using the provided methods in CAMERA (x$instrument_raw, x$instrument_maxz, x$instrument_susie, x$instrument_paintor). Default is x$instrument_raw.}
      \item{\code{p_exp}}{Statistical threshold to determine significance. Default is "bonferroni", which is eqaul to 0.05/number of the instruments.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Data frame in x$instrument_outcome
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-make_outcome_local"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-make_outcome_local}{}}}
\subsection{\code{CAMERA$make_outcome_local()}}{
  The function extracts summary statistics of given a list of instruments and the outcomes from local data
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$make_outcome_local(
  exp = self$instrument_raw,
  out = self$instrument_outcome_regions,
  p_exp = 0.05/nrow(exp)
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{exp}}{Intsruments for the exposure that are selected by using the provided methods in CAMERA (x$instrument_raw, x$instrument_fema, x$instrument_susie, x$instrument_paintor). Default is x$instrument_raw.}
      \item{\code{out}}{Outcome data for each instrument region. Default is x$instrument_outcome_regions.}
      \item{\code{p_exp}}{Statistical threshold to determine significance. Default is "bonferroni", which is eqaul to 0.05/number of the instruments.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Data frame in x$instrument_outcome
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-harmonise"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-harmonise}{}}}
\subsection{\code{CAMERA$harmonise()}}{
  This function harmonises the alleles and effects between the exposure and outcome.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$harmonise(exp = self$instrument_raw, out = self$instrument_outcome)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{exp}}{Intsruments for the exposure that are selected by using the provided methods in CAMERA (x$instrument_raw, x$instrument_fema, x$instrument_susie, x$instrument_paintor). Default is x$instrument_raw.}
      \item{\code{out}}{Intsruments for the outcome by using \code{make_outcome_data()}. Default is x$instrument_outcome.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Data frame in x$harmonised_dat_sem
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-set_summary"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-set_summary}{}}}
\subsection{\code{CAMERA$set_summary()}}{
  Generate summary of exposure, outcome and population metadata
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$set_summary()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    data frame
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-pleiotropy"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-pleiotropy}{}}}
\subsection{\code{CAMERA$pleiotropy()}}{
  Estimate similarity of horizontal pleiotropy across ancestries.
For each ancestry, identify outliers in the MR analysis based on per-variatn Q statistics. Then estimate the deviation from the main estimates for all outliers across all ancestries. Finally, determine if the pleiotropy deviation is consistent across all ancestries
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$pleiotropy(harmonised_dat = self$harmonised_dat, mrres = self$mrres)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{harmonised_dat}}{Outcome from \code{harmonise} function}
      \item{\code{mrres}}{Outcome from \code{cross_estimate}}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \itemize{
\item data frame of outliers \code{pleiotropy_outliers}
\item data frame of heterogeneity for each outlier / ancestry combination \code{pleiotropy_Q_outliers}
\item data frame of agreement of outlier effects \code{pleiotropy_agreement}
}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-plot_pleiotropy"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-plot_pleiotropy}{}}}
\subsection{\code{CAMERA$plot_pleiotropy()}}{
  Plot pleiotropy results
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$plot_pleiotropy(dat = self$pleiotropy_outliers)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dat}}{Output from \code{pleiotropy} - \code{pleiotropy_outliers}}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    plot
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-plot_pleiotropy_heterogeneity"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-plot_pleiotropy_heterogeneity}{}}}
\subsection{\code{CAMERA$plot_pleiotropy_heterogeneity()}}{
  Plot pleiotropy results per variant
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$plot_pleiotropy_heterogeneity(
  dat = self$pleiotropy_Q_outliers,
  pthresh = 0.05
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dat}}{Output from \code{pleiotropy} - \code{pleiotropy_Q_outliers}}
      \item{\code{pthresh}}{p-value from Q statistic for inclusion in plots}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    plot
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-perform_basic_sem"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-perform_basic_sem}{}}}
\subsection{\code{CAMERA$perform_basic_sem()}}{
  Perform basic SEM analysis of the joint estimates in multiple ancestries.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$perform_basic_sem(harmonised_dat = self$harmonised_dat_sem)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{harmonised_dat}}{Harmonised dataset obtained by using \code{harmonised_dat()}}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Table of the SEM results. Summary of the results available in x$sem_result.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-runsem"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-runsem}{}}}
\subsection{\code{CAMERA$runsem()}}{
  Fit an SEM model using lavaan
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$runsem(model, data, modname)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{model}}{lavaan model syntax}
      \item{\code{data}}{Data frame of the data for the model}
      \item{\code{modname}}{Name of the model used to label the results}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Data frame of the SEM estimates for each population
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-standardise_data"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-standardise_data}{}}}
\subsection{\code{CAMERA$standardise_data()}}{
  This function standardises the betas and SEs for the instruments-exposure/outcome associations when unit information is not matched across the populations. In case of large differences in genetic effects observed between the populations (e.g. due to sample size difference), the function scales the betas and SEs.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$standardise_data(
  dat = self$instrument_raw,
  standardise_unit = FALSE,
  standardise_scale = FALSE,
  scaling_method = "simple_mode"
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dat}}{Instruments for the exposure that are selected by using the provided methods in CAMERA (\code{x$instrument_raw}, \code{x$instrument_maxz}, \code{x$instrument_susie}, \code{x$instrument_paintor}). Default is \code{x$instrument_raw}.}
      \item{\code{dat}}{A data frame containing the harmonised data. It should have the columns \code{beta.y}, \code{beta.x}, \code{se.y}, and \code{pops}. If not provided, the method uses the \code{harmonised_dat} attribute of the \code{CAMERA} object.}
      \item{\code{dat}}{Genomic regions identified by using \code{x$extract_instrument_regions()}}
      \item{\code{dat}}{Harmonised data. Default is \code{x$harmonised_dat}.}
      \item{\code{standardise_unit}}{Use this option if unit information is not matched.}
      \item{\code{standardise_scale}}{Use this option if genetic effects are substantially different due to study power.}
      \item{\code{scaling_method}}{Choose the methods to obtain scaling units (MR estimates of exposure 1 and exposure 2 or outcome 1 and outcome 2). Default is \code{"simple_mode"}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-clone}{}}}
\subsection{\code{CAMERA$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CAMERA$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
