% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundQuantiles.R
\name{compoundLibraries}
\alias{compoundLibraries}
\title{The supported compound databases}
\usage{
compoundLibraries()
}
\value{
Vector of supported compound databases
}
\description{
Returns a set of supported compound databases
}
\examples{
compoundLibraries()
}
\author{
Hendrik Treutler
}
