% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundQuantiles.R
\name{compoundQuantiles}
\alias{compoundQuantiles}
\title{compoundQuantiles constructor}
\usage{
compoundQuantiles(compoundLibrary = "kegg", massWindowSize = 50)
}
\arguments{
\item{compoundLibrary}{the database; see compoundLibraries() for a list of supported databases}

\item{massWindowSize}{the mass window size for grouping compounds; see massWindowSizes(compoundLibrary = "kegg") for a list of supported databases for e.g. the database kegg}
}
\value{
the compoundQuantiles object
}
\description{
constructor of class compoundQuantiles
}
\examples{
cpObj <- compoundQuantiles()
}
\author{
Hendrik Treutler
}
