% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgxe.R
\name{egger_bootstrap}
\alias{egger_bootstrap}
\title{Perform MR GxE}
\usage{
egger_bootstrap(b_gx, se_gx, b_gy, se_gy, nboot = 1000)
}
\arguments{
\item{b_gx}{Vector of instrument-exposure associations, one for each sub group}

\item{se_gx}{Vector of standard errors to b_gx}

\item{b_gy}{Vector of instrument-outcome associations, one for each sub group}

\item{se_gy}{Vector of standard errors for b_gy}

\item{nboot}{Number of bootstraps. Default=1000}
}
\value{
List
\itemize{
\item a = intercept estimate (pleiotropy)
\item b = slope estimate (b_iv effect)
\item a_se = standard error of intercept
\item b_se = standard error of slope
\item a_pval = p-value of intercept estimate
\item b_pval = p-value of slope estimate
\item a_mean = mean value of intercept from bootstraps
\item b_mean = mean value of slope estimates from bootstraps
}
}
\description{
For a single variant estiamted in different sub groups.
}
\details{
Estimate the degree of pleiotropy using MR GxE. This method uses a negative control type approach based on an assumption that the instrument-exposure association is uncorrelated with the pleiotropic effect. Therefore, as the instrument-exposure association reduces in magnitude, the effect on the outcome will reduce towards an intercept term which represents the pleiotropic effect.

Standard errors are obtained from parametric bootstrap
}
