% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundQuantiles.R
\name{example}
\alias{example}
\title{Runs an example}
\usage{
example()
}
\value{
void
}
\description{
Runs an example for the usage of package compoundQuantiles
}
\examples{
example()
}
\author{
Hendrik Treutler
}
