% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_findIsotopes.R
\name{findIsotopePeaks}
\alias{findIsotopePeaks}
\title{Find which peaks are isotopes of each other}
\usage{
findIsotopePeaks(
  mz,
  int,
  ipeak = seq_along(mz),
  ppm = 5,
  tolerance = 0,
  maxiso = 4L,
  maxcharge = 3L,
  isomatrix = calcIsotopeMatrix(maxiso),
  minfrac = 0.5,
  filter = TRUE
)
}
\arguments{
\item{mz}{\code{numeric} with the m/z values of the peaks.}

\item{int}{\code{numeric} or \code{matrix} with intensity values for the individual
peaks. A \code{matrix} (with number of rows equal to the number of peaks)
can be provided if intensities are available across several samples.}

\item{ipeak}{optional \code{integer}, same length than \code{mz} providing the index
of the peaks in the original data set. These indices are returned as
indices of original and isotope peaks in the result matrix. It defaults
to a sequence from 1 to length of \code{mz}.}

\item{ppm}{\code{numeric(1)} with the relative acceptable difference in m/z
values. Defaults to \code{ppm = 5}.}

\item{tolerance}{\code{numeric(1)} specifying an absolute difference in m/z
values. Defaults to \code{tolerance = 0}.}

\item{maxiso}{\code{integer(1)} with the maximal number of isotopes that should
be searched for.}

\item{maxcharge}{\code{integer(1)} with the maximal charge of the expected
isotopets.}

\item{isomatrix}{\code{matrix} with isotope relationship definitions, as the
one returned by \code{\link[=calcIsotopeMatrix]{calcIsotopeMatrix()}}.}

\item{minfrac}{\code{numeric(1)} defining the proportion of samples which have
to show the correct C12/C13 ratio, if \code{int} is a \code{matrix} with intensity
values for multiple samples.}

\item{filter}{\code{logical(1)} whether the expected intensity relationship for
C12/C13 filter should be applied.}
}
\value{
\code{matrix} with 4 columns:
\itemize{
\item \code{"mpeak"}: index of the peak.
\item \code{"isopeak"}: index of the peak identified as an isotope of \code{"mpeak"}.
\item \code{"iso"}: the isotope.
\item \code{"charge"}: the charge of the isotope.
}
}
\description{
\code{findIsotopePeaks} compares diffrences between m/z of the provided peaks to
determine which peaks could be isotopes of each other considering also
the relationship between their intensities.
}
