% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{fixed_effects_meta_analysis}
\alias{fixed_effects_meta_analysis}
\title{Perform fixed effects meta analysis for one association}
\usage{
fixed_effects_meta_analysis(beta_vec, se_vec, infl = 10000)
}
\arguments{
\item{infl}{Inflation factor - how much larger is the estimate than the estimate of the tightest SE - for use in removing unreliable estimates}
}
\value{
list of results
}
\description{
Perform fixed effects meta analysis for one association
}
