% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_selection.R
\name{fixed_effects_meta_analysis_fast}
\alias{fixed_effects_meta_analysis_fast}
\title{Fixed effects meta analysis vectorised across multiple SNPs}
\usage{
fixed_effects_meta_analysis_fast(beta_mat, se_mat)
}
\arguments{
\item{beta_mat}{Matrix of betas - rows are SNPs, columns are studies}

\item{se_mat}{Matrix of SEs - rows are SNPs, columns are studies}
}
\value{
list of meta analysis betas and SEs
}
\description{
Assumes effects across studies are all on the same scale
}
