% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{identify_blownup_estimates}
\alias{identify_blownup_estimates}
\title{Identify blown up estimates}
\usage{
identify_blownup_estimates(b, se, infl)
}
\arguments{
\item{b}{Vector of betas}

\item{se}{Vector of SEs}

\item{infl}{Inflation factor - how much larger is the estimate than the estimate of the tightest SE}
}
\value{
index of betas to remove
}
\description{
Sometimes estimates appear unstable. They are likely unreliable and best to not use for heterogeneity analyses etc.
}
