% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundQuantiles.R
\name{massWindowSizes}
\alias{massWindowSizes}
\title{The supported mass window sizes}
\usage{
massWindowSizes(libraryName = "kegg")
}
\arguments{
\item{libraryName}{The compound database}
}
\value{
Vector of supported mass window sizes
}
\description{
Returns the set of supported mass window sizes for the given compound database
}
\examples{
massWindowSizes()
}
\author{
Hendrik Treutler
}
