% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAncestrySummarySet.R
\name{MultiAncestrySummarySet}
\alias{MultiAncestrySummarySet}
\title{R6 class for MR.TRAM}
\description{
A simple wrapper function.
Using a summary set, identify set of instruments for the traits, and peform SEM MR to test the association across the population.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-import}{\code{MultiAncestrySummarySet$import()}}
\item \href{#method-new}{\code{MultiAncestrySummarySet$new()}}
\item \href{#method-extract_instruments}{\code{MultiAncestrySummarySet$extract_instruments()}}
\item \href{#method-check_instruments}{\code{MultiAncestrySummarySet$check_instruments()}}
\item \href{#method-extract_instrument_regions}{\code{MultiAncestrySummarySet$extract_instrument_regions()}}
\item \href{#method-estimate_instrument_specificity}{\code{MultiAncestrySummarySet$estimate_instrument_specificity()}}
\item \href{#method-prop_overlap}{\code{MultiAncestrySummarySet$prop_overlap()}}
\item \href{#method-scan_regional_instruments}{\code{MultiAncestrySummarySet$scan_regional_instruments()}}
\item \href{#method-plot_regional_instruments}{\code{MultiAncestrySummarySet$plot_regional_instruments()}}
\item \href{#method-regional_ld_matrices}{\code{MultiAncestrySummarySet$regional_ld_matrices()}}
\item \href{#method-susie_finemap_regions}{\code{MultiAncestrySummarySet$susie_finemap_regions()}}
\item \href{#method-paintor_finemap_regions}{\code{MultiAncestrySummarySet$paintor_finemap_regions()}}
\item \href{#method-extract_outcome_data}{\code{MultiAncestrySummarySet$extract_outcome_data()}}
\item \href{#method-perform_basic_sem}{\code{MultiAncestrySummarySet$perform_basic_sem()}}
\item \href{#method-clone}{\code{MultiAncestrySummarySet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import"></a>}}
\if{latex}{\out{\hypertarget{method-import}{}}}
\subsection{Method \code{import()}}{
Migrate the results from a previous MultiAncestrySummarySet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$import(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new dataset and initialise an R interface
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$new(
  exposure_ids = NULL,
  outcome_ids = NULL,
  pops = NULL,
  bfiles = NULL,
  plink = NULL,
  radius = NULL,
  clump_pop = NULL,
  x = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exposure_ids}}{IDs for the exposures from each population.}

\item{\code{outcome_ids}}{IDs for the outcomes from each population.}

\item{\code{pops}}{Ancestry information; AFR, AMR, EUR, EAS, SAS}

\item{\code{bfiles}}{Location of LD reference file (Download from: http://fileserve.mrcieu.ac.uk/ld/1kg.v3.tgz)}

\item{\code{plink}}{Location of executable plink (ver.1.90)}

\item{\code{radius}}{Set a range of the region to look for}

\item{\code{clump_pop}}{insert}

\item{\code{x}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-extract_instruments}{}}}
\subsection{Method \code{extract_instruments()}}{
Identifies the instruments for the exposure
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$extract_instruments(
  exposure_ids = self$exposure_ids,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exposure_ids}}{ID for the exposure. Default is x$exposure_ids.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-check_instruments}{}}}
\subsection{Method \code{check_instruments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$check_instruments(ids = self$exposure_ids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract_instrument_regions"></a>}}
\if{latex}{\out{\hypertarget{method-extract_instrument_regions}{}}}
\subsection{Method \code{extract_instrument_regions()}}{
Extract a region around each instrument obtained from \code{x$extract_instruments()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$extract_instrument_regions(
  radius = self$radius,
  instrument_raw = self$instrument_raw,
  exposure_ids = self$exposure_ids
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{radius}}{Set a range of the region to search}

\item{\code{instrument_raw}}{A set of instruments obtained from \code{x$extract_instruments()}}

\item{\code{exposure_ids}}{ID for the exposure. Default is x$exposure_ids.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimate_instrument_specificity"></a>}}
\if{latex}{\out{\hypertarget{method-estimate_instrument_specificity}{}}}
\subsection{Method \code{estimate_instrument_specificity()}}{
Estimate what fraction of the instuments is expected to be replicated under the hypothesis that the effect estimates are the same
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$estimate_instrument_specificity(
  instrument,
  alpha = "bonferroni"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instrument}}{A set of instruments obtained from \code{x$extract_instruments()} or \code{scan_regional_instruments()}}

\item{\code{alpha}}{Significance level. Default is 0.05/number of SNPs (Bonferroni)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prop_overlap"></a>}}
\if{latex}{\out{\hypertarget{method-prop_overlap}{}}}
\subsection{Method \code{prop_overlap()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$prop_overlap(b_disc, b_rep, se_disc, se_rep, alpha)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b_disc}}{insert}

\item{\code{b_rep}}{insert}

\item{\code{se_disc}}{insert}

\item{\code{se_rep}}{insert}

\item{\code{alpha}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scan_regional_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-scan_regional_instruments}{}}}
\subsection{Method \code{scan_regional_instruments()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$scan_regional_instruments(
  instrument_raw = self$instrument_raw,
  instrument_regions = self$instrument_regions
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instrument_raw}}{insert}

\item{\code{instrument_regions}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_regional_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-plot_regional_instruments}{}}}
\subsection{Method \code{plot_regional_instruments()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$plot_regional_instruments(
  region = 1:min(10, nrow(instruments)),
  instrument_region_zscores = self$instrument_region_zscores,
  instruments = self$instrument_raw
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{insert}

\item{\code{instrument_region_zscores}}{insert}

\item{\code{instruments}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-regional_ld_matrices"></a>}}
\if{latex}{\out{\hypertarget{method-regional_ld_matrices}{}}}
\subsection{Method \code{regional_ld_matrices()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$regional_ld_matrices(
  instrument_regions = self$instrument_regions,
  bfiles = self$bfiles,
  pops = self$pops,
  plink = self$plink
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instrument_regions}}{insert}

\item{\code{bfiles}}{insert}

\item{\code{pops}}{insert}

\item{\code{plink}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-susie_finemap_regions"></a>}}
\if{latex}{\out{\hypertarget{method-susie_finemap_regions}{}}}
\subsection{Method \code{susie_finemap_regions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$susie_finemap_regions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-paintor_finemap_regions"></a>}}
\if{latex}{\out{\hypertarget{method-paintor_finemap_regions}{}}}
\subsection{Method \code{paintor_finemap_regions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$paintor_finemap_regions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract_outcome_data"></a>}}
\if{latex}{\out{\hypertarget{method-extract_outcome_data}{}}}
\subsection{Method \code{extract_outcome_data()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$extract_outcome_data(
  exp = self$instrument_raw,
  standardise = FALSE,
  p_exp = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exp}}{insert}

\item{\code{p_exp}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-perform_basic_sem"></a>}}
\if{latex}{\out{\hypertarget{method-perform_basic_sem}{}}}
\subsection{Method \code{perform_basic_sem()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$perform_basic_sem(
  harmonised_dat = self$harmonised_dat_sem
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{harmonised_dat}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
