% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAncestrySummarySet.R
\name{plot_MsCAVIAR}
\alias{plot_MsCAVIAR}
\title{Plot MsCAVIAR results}
\usage{
plot_MsCAVIAR(res)
}
\arguments{
\item{res}{Output from run_MsCAVIAR}
}
\value{
plot
}
\description{
Plot MsCAVIAR results
}
