% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAncestrySummarySet.R
\name{plot_PAINTOR}
\alias{plot_PAINTOR}
\title{Plot PAINTOR results}
\usage{
plot_PAINTOR(res)
}
\arguments{
\item{res}{Output from run_PAINTOR}
}
\value{
plot
}
\description{
Plot PAINTOR results
}
