% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_instruments.R
\name{prop_overlap}
\alias{prop_overlap}
\title{Estimate expected vs observed replication of effects between discovery and replication datasets}
\usage{
prop_overlap(b_disc, b_rep, se_disc, se_rep, alpha)
}
\arguments{
\item{b_disc}{Vector of discovery betas}

\item{b_rep}{Vector of replication betas}

\item{se_disc}{Vector of discovery standard errors}

\item{se_rep}{Vector of replication standard errors}

\item{alpha}{Nominal replication significance threshold}
}
\value{
List of results
\itemize{
\item res: aggregate expected replication rate vs observed replication rate
\item variants: per variant expected replication rates
}
}
\description{
Taken from Okbay et al 2016. Under the assumption that all discovery effects are unbiased, what fraction of associations would replicate in the replication dataset, given the differential power of the discovery and replication datasets.
Uses standard error of the replication dataset to account for differences in sample size and distribution of independent variable
}
