% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAncestrySummarySet.R
\name{run_MsCAVIAR}
\alias{run_MsCAVIAR}
\title{Analyse regional data with MsCAVIAR}
\usage{
run_MsCAVIAR(regiondata, n, MsCAVIAR = "MsCAVIAR", workdir = tempdir())
}
\arguments{
\item{regiondata}{Output from extract_regional_data}

\item{MsCAVIAR}{Path to MsCAVIAR executable. Default="MsCAVIAR"}

\item{workdir}{Working directory. Default=tempdir()}
}
\value{
Results table with posterior inclusion probabilities
}
\description{
Analyse regional data with MsCAVIAR
}
