% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAncestrySummarySet.R
\name{run_PAINTOR}
\alias{run_PAINTOR}
\title{Analyse regional data with PAINTOR}
\usage{
run_PAINTOR(regiondata, PAINTOR = "PAINTOR", workdir = tempdir())
}
\arguments{
\item{regiondata}{Output from extract_regional_data}

\item{PAINTOR}{Path to PAINTOR executable. Default="PAINTOR"}

\item{workdir}{Working directory. Default=tempdir()}
}
\value{
Results table with posterior inclusion probabilities
}
\description{
Analyse regional data with PAINTOR
}
