% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAMERA.R
\name{CAMERA}
\alias{CAMERA}
\title{R6 class for CAMERA}
\description{
A simple wrapper function.
Using a summary set, identify set of instruments for the traits, and peform SEM MR to test the association across the population.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-import}{\code{CAMERA$import()}}
\item \href{#method-new}{\code{CAMERA$new()}}
\item \href{#method-check_phenotypes}{\code{CAMERA$check_phenotypes()}}
\item \href{#method-extract_instruments}{\code{CAMERA$extract_instruments()}}
\item \href{#method-extract_instrument_regions}{\code{CAMERA$extract_instrument_regions()}}
\item \href{#method-scan_regional_instruments}{\code{CAMERA$scan_regional_instruments()}}
\item \href{#method-plot_regional_instruments}{\code{CAMERA$plot_regional_instruments()}}
\item \href{#method-regional_ld_matrices}{\code{CAMERA$regional_ld_matrices()}}
\item \href{#method-replication_evaluation}{\code{CAMERA$replication_evaluation()}}
\item \href{#method-susie_finemap_regions}{\code{CAMERA$susie_finemap_regions()}}
\item \href{#method-paintor_finemap_regions}{\code{CAMERA$paintor_finemap_regions()}}
\item \href{#method-MsCAVIAR_finemap_regions}{\code{CAMERA$MsCAVIAR_finemap_regions()}}
\item \href{#method-instrument_heterogeneity}{\code{CAMERA$instrument_heterogeneity()}}
\item \href{#method-standardise_data}{\code{CAMERA$standardise_data()}}
\item \href{#method-estimate_instrument_specificity}{\code{CAMERA$estimate_instrument_specificity()}}
\item \href{#method-make_outcome_data}{\code{CAMERA$make_outcome_data()}}
\item \href{#method-harmonised_dat}{\code{CAMERA$harmonised_dat()}}
\item \href{#method-perform_basic_sem}{\code{CAMERA$perform_basic_sem()}}
\item \href{#method-clone}{\code{CAMERA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import"></a>}}
\if{latex}{\out{\hypertarget{method-import}{}}}
\subsection{Method \code{import()}}{
Migrate the results from a previous CAMERA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$import(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new dataset and initialise an R interface
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$new(
  exposure_ids = NULL,
  outcome_ids = NULL,
  pops = NULL,
  bfiles = NULL,
  plink = NULL,
  radius = NULL,
  clump_pop = NULL,
  x = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exposure_ids}}{IDs for the exposures from each population.}

\item{\code{outcome_ids}}{IDs for the outcomes from each population.}

\item{\code{pops}}{Ancestry information; AFR, AMR, EUR, EAS, SAS}

\item{\code{bfiles}}{Location of LD reference file (Download from: http://fileserve.mrcieu.ac.uk/ld/1kg.v3.tgz)}

\item{\code{plink}}{Location of executable plink (ver.1.90)}

\item{\code{radius}}{Set a range of the region to look for}

\item{\code{clump_pop}}{insert}

\item{\code{x}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_phenotypes"></a>}}
\if{latex}{\out{\hypertarget{method-check_phenotypes}{}}}
\subsection{Method \code{check_phenotypes()}}{
Evaluate instrument associations between populations. The function checks whether 1) the instruments (chosen summary statistics IDs) can be used for the further steps and 2) the units are comparable between populations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$check_phenotypes(ids = self$exposure_ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{ID for the exposure or the outcome. Default is x$exposure_ids.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-extract_instruments}{}}}
\subsection{Method \code{extract_instruments()}}{
Identifies the instruments for the exposure
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$extract_instruments(exposure_ids = self$exposure_ids, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exposure_ids}}{ID for the exposure. Default is x$exposure_ids.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract_instrument_regions"></a>}}
\if{latex}{\out{\hypertarget{method-extract_instrument_regions}{}}}
\subsection{Method \code{extract_instrument_regions()}}{
Extract a region around each instrument obtained from \code{x$extract_instruments()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$extract_instrument_regions(
  radius = self$radius,
  instrument_raw = self$instrument_raw,
  exposure_ids = self$exposure_ids
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{radius}}{Set a range of the region to search}

\item{\code{instrument_raw}}{A set of instruments obtained from \code{x$extract_instruments()}}

\item{\code{exposure_ids}}{ID for the exposure. Default is x$exposure_ids.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scan_regional_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-scan_regional_instruments}{}}}
\subsection{Method \code{scan_regional_instruments()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$scan_regional_instruments(
  instrument_raw = self$instrument_raw,
  instrument_regions = self$instrument_regions
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instrument_raw}}{insert}

\item{\code{instrument_regions}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_regional_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-plot_regional_instruments}{}}}
\subsection{Method \code{plot_regional_instruments()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$plot_regional_instruments(
  instrument_region_zscores = self$instrument_region_zscores,
  instruments = self$instrument_raw,
  region = 1:min(10, nrow(instruments)),
  comparison = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instrument_region_zscores}}{insert}

\item{\code{instruments}}{insert}

\item{\code{region}}{insert}

\item{\code{comparison}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-regional_ld_matrices"></a>}}
\if{latex}{\out{\hypertarget{method-regional_ld_matrices}{}}}
\subsection{Method \code{regional_ld_matrices()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$regional_ld_matrices(
  instrument_regions = self$instrument_regions,
  bfiles = self$bfiles,
  pops = self$pops,
  plink = self$plink
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instrument_regions}}{insert}

\item{\code{bfiles}}{insert}

\item{\code{pops}}{insert}

\item{\code{plink}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replication_evaluation"></a>}}
\if{latex}{\out{\hypertarget{method-replication_evaluation}{}}}
\subsection{Method \code{replication_evaluation()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$replication_evaluation(
  instrument = self$instrument_raw,
  ld = self$ld_matrices
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instrument}}{insert}

\item{\code{ld}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-susie_finemap_regions"></a>}}
\if{latex}{\out{\hypertarget{method-susie_finemap_regions}{}}}
\subsection{Method \code{susie_finemap_regions()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$susie_finemap_regions(
  dat = self$instrument_regions,
  ld = self$ld_matrices
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{insert}

\item{\code{ld}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-paintor_finemap_regions"></a>}}
\if{latex}{\out{\hypertarget{method-paintor_finemap_regions}{}}}
\subsection{Method \code{paintor_finemap_regions()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$paintor_finemap_regions(
  region = self$instrument_regions,
  ld = self$ld_matrices,
  PAINTOR = "PAINTOR",
  workdir = tempdir()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{Output from extract_regional_data}

\item{\code{ld}}{insert}

\item{\code{PAINTOR}}{Path to PAINTOR executable. Default="PAINTOR"}

\item{\code{workdir}}{Working directory. Default=tempdir()}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MsCAVIAR_finemap_regions"></a>}}
\if{latex}{\out{\hypertarget{method-MsCAVIAR_finemap_regions}{}}}
\subsection{Method \code{MsCAVIAR_finemap_regions()}}{
Analyse regional data with MsCAVIAR
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$MsCAVIAR_finemap_regions(
  region = self$instrument_regions,
  ld = self$ld_matrices,
  MsCAVIAR = "MsCAVIAR",
  workdir = tempdir()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{Output from extract_regional_data}

\item{\code{ld}}{insert}

\item{\code{MsCAVIAR}}{Path to MsCAVIAR executable. Default="MsCAVIAR"}

\item{\code{workdir}}{Working directory. Default=tempdir()}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-instrument_heterogeneity"></a>}}
\if{latex}{\out{\hypertarget{method-instrument_heterogeneity}{}}}
\subsection{Method \code{instrument_heterogeneity()}}{
Evaluate heterogeneity of identified instrument-trait associations across populations, where the instruments were identified using "Raw", "MaxZ", or fine-mapping (Susie, PAINTOR) methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$instrument_heterogeneity(
  instrument = self$instrument_raw,
  alpha = "bonferroni"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instrument}}{Intsruments for the exposure that are selected the provided methods (x$instrument_raw, x$instrument_maxz, x$instrument_susie, x$instrument_paintor). Default is x$instrument_raw.}

\item{\code{alpha}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standardise_data"></a>}}
\if{latex}{\out{\hypertarget{method-standardise_data}{}}}
\subsection{Method \code{standardise_data()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$standardise_data(
  dat = self$instrument_raw,
  standardise_unit = FALSE,
  standardise_scale = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{insert}

\item{\code{standardise_unit}}{insert}

\item{\code{standardise_scale}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimate_instrument_specificity"></a>}}
\if{latex}{\out{\hypertarget{method-estimate_instrument_specificity}{}}}
\subsection{Method \code{estimate_instrument_specificity()}}{
Estimate what fraction of the instuments is expected to be replicated under the hypothesis that the effect estimates are the same
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$estimate_instrument_specificity(
  instrument,
  alpha = "bonferroni",
  winnerscurse = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instrument}}{A set of instruments obtained from \code{x$extract_instruments()} or \code{scan_regional_instruments()}}

\item{\code{alpha}}{Significance level. Default is 0.05/number of SNPs (Bonferroni)}

\item{\code{winnerscurse}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_outcome_data"></a>}}
\if{latex}{\out{\hypertarget{method-make_outcome_data}{}}}
\subsection{Method \code{make_outcome_data()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$make_outcome_data(exp = self$instrument_raw, p_exp = 0.05/nrow(exp))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exp}}{insert}

\item{\code{p_exp}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-harmonised_dat"></a>}}
\if{latex}{\out{\hypertarget{method-harmonised_dat}{}}}
\subsection{Method \code{harmonised_dat()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$harmonised_dat(exp = self$instrument_raw, out = self$instrument_outcome)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exp}}{insert}

\item{\code{out}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-perform_basic_sem"></a>}}
\if{latex}{\out{\hypertarget{method-perform_basic_sem}{}}}
\subsection{Method \code{perform_basic_sem()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$perform_basic_sem(harmonised_dat = self$harmonised_dat_sem)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{harmonised_dat}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
