% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00CAMERA.R
\name{CAMERA}
\alias{CAMERA}
\title{R6 class for CAMERA}
\description{
A simple wrapper function.
Using a summary set, identify set of instruments for the traits, and peform SEM MR to test the association across the population.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CAMERA-import}{\code{CAMERA$import()}}
\item \href{#method-CAMERA-assign}{\code{CAMERA$assign()}}
\item \href{#method-CAMERA-import_from_local}{\code{CAMERA$import_from_local()}}
\item \href{#method-CAMERA-new}{\code{CAMERA$new()}}
\item \href{#method-CAMERA-instrument_heterogeneity}{\code{CAMERA$instrument_heterogeneity()}}
\item \href{#method-CAMERA-estimate_instrument_specificity}{\code{CAMERA$estimate_instrument_specificity()}}
\item \href{#method-CAMERA-replication_evaluation}{\code{CAMERA$replication_evaluation()}}
\item \href{#method-CAMERA-check_phenotypes}{\code{CAMERA$check_phenotypes()}}
\item \href{#method-CAMERA-cross_estimate}{\code{CAMERA$cross_estimate()}}
\item \href{#method-CAMERA-plot_cross_estimate}{\code{CAMERA$plot_cross_estimate()}}
\item \href{#method-CAMERA-extract_instruments}{\code{CAMERA$extract_instruments()}}
\item \href{#method-CAMERA-extract_instrument_regions}{\code{CAMERA$extract_instrument_regions()}}
\item \href{#method-CAMERA-scan_regional_instruments}{\code{CAMERA$scan_regional_instruments()}}
\item \href{#method-CAMERA-plot_regional_instruments_maxz}{\code{CAMERA$plot_regional_instruments_maxz()}}
\item \href{#method-CAMERA-regional_ld_matrices}{\code{CAMERA$regional_ld_matrices()}}
\item \href{#method-CAMERA-susie_finemap_regions}{\code{CAMERA$susie_finemap_regions()}}
\item \href{#method-CAMERA-paintor_finemap_regions}{\code{CAMERA$paintor_finemap_regions()}}
\item \href{#method-CAMERA-MsCAVIAR_finemap_regions}{\code{CAMERA$MsCAVIAR_finemap_regions()}}
\item \href{#method-CAMERA-fema_regional_instruments}{\code{CAMERA$fema_regional_instruments()}}
\item \href{#method-CAMERA-plot_regional_instruments}{\code{CAMERA$plot_regional_instruments()}}
\item \href{#method-CAMERA-get_metadata}{\code{CAMERA$get_metadata()}}
\item \href{#method-CAMERA-estimate_instrument_heterogeneity_per_variant}{\code{CAMERA$estimate_instrument_heterogeneity_per_variant()}}
\item \href{#method-CAMERA-mrgxe}{\code{CAMERA$mrgxe()}}
\item \href{#method-CAMERA-mrgxe_plot}{\code{CAMERA$mrgxe_plot()}}
\item \href{#method-CAMERA-mrgxe_plot_variant}{\code{CAMERA$mrgxe_plot_variant()}}
\item \href{#method-CAMERA-make_outcome_data}{\code{CAMERA$make_outcome_data()}}
\item \href{#method-CAMERA-make_outcome_local}{\code{CAMERA$make_outcome_local()}}
\item \href{#method-CAMERA-harmonise}{\code{CAMERA$harmonise()}}
\item \href{#method-CAMERA-set_summary}{\code{CAMERA$set_summary()}}
\item \href{#method-CAMERA-pleiotropy}{\code{CAMERA$pleiotropy()}}
\item \href{#method-CAMERA-plot_pleiotropy}{\code{CAMERA$plot_pleiotropy()}}
\item \href{#method-CAMERA-plot_pleiotropy_heterogeneity}{\code{CAMERA$plot_pleiotropy_heterogeneity()}}
\item \href{#method-CAMERA-perform_basic_sem}{\code{CAMERA$perform_basic_sem()}}
\item \href{#method-CAMERA-runsem}{\code{CAMERA$runsem()}}
\item \href{#method-CAMERA-standardise_data}{\code{CAMERA$standardise_data()}}
\item \href{#method-CAMERA-clone}{\code{CAMERA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-import"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-import}{}}}
\subsection{Method \code{import()}}{
Migrate the results from a previous CAMERA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$import(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{R6 Environment created for CAMERA. Default = x}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-assign"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-assign}{}}}
\subsection{Method \code{assign()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$assign(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-import_from_local"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-import_from_local}{}}}
\subsection{Method \code{import_from_local()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$import_from_local(
  instrument_raw,
  instrument_outcome,
  instrument_regions,
  instrument_outcome_regions,
  exposure_ids,
  outcome_ids,
  pops,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-new"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-new}{}}}
\subsection{Method \code{new()}}{
Create a new dataset and initialise an R interface
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$new(
  exposure_ids = NULL,
  outcome_ids = NULL,
  pops = NULL,
  bfiles = NULL,
  plink = NULL,
  radius = NULL,
  clump_pop = NULL,
  x = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exposure_ids}}{Exposures IDs obtained from IEU GWAS database (https://gwas.mrcieu.ac.uk/) for each population}

\item{\code{outcome_ids}}{Outcome IDs obtained from IEU GWAS database (https://gwas.mrcieu.ac.uk/) for each population}

\item{\code{pops}}{Ancestry information for each population (i.e. AFR, AMR, EUR, EAS, SAS)}

\item{\code{bfiles}}{Locations of LD reference files for each population (Download from: http://fileserve.mrcieu.ac.uk/ld/1kg.v3.tgz)}

\item{\code{plink}}{Location of executable plink (ver.1.90 is recommended)}

\item{\code{radius}}{Genomic window size to extract SNPs}

\item{\code{clump_pop}}{Reference population for clumping}

\item{\code{x}}{Import data where available}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-instrument_heterogeneity"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-instrument_heterogeneity}{}}}
\subsection{Method \code{instrument_heterogeneity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$instrument_heterogeneity(
  instrument = self$instrument_raw,
  alpha = "bonferroni",
  method = "ivw",
  outlier_removal = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-estimate_instrument_specificity"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-estimate_instrument_specificity}{}}}
\subsection{Method \code{estimate_instrument_specificity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$estimate_instrument_specificity(
  instrument,
  alpha = "bonferroni",
  winnerscurse = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-replication_evaluation"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-replication_evaluation}{}}}
\subsection{Method \code{replication_evaluation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$replication_evaluation(
  instrument = self$instrument_raw,
  ld = self$ld_matrices
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-check_phenotypes"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-check_phenotypes}{}}}
\subsection{Method \code{check_phenotypes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$check_phenotypes(ids = self$exposure_ids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-cross_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-cross_estimate}{}}}
\subsection{Method \code{cross_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$cross_estimate(dat = self$harmonised_dat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-plot_cross_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-plot_cross_estimate}{}}}
\subsection{Method \code{plot_cross_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$plot_cross_estimate(res = self$mrres, qj_alpha = 0.05)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-extract_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-extract_instruments}{}}}
\subsection{Method \code{extract_instruments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$extract_instruments(exposure_ids = self$exposure_ids, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-extract_instrument_regions"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-extract_instrument_regions}{}}}
\subsection{Method \code{extract_instrument_regions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$extract_instrument_regions(
  radius = self$radius,
  instrument_raw = self$instrument_raw,
  exposure_ids = self$exposure_ids
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-scan_regional_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-scan_regional_instruments}{}}}
\subsection{Method \code{scan_regional_instruments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$scan_regional_instruments(
  instrument_raw = self$instrument_raw,
  instrument_regions = self$instrument_regions
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-plot_regional_instruments_maxz"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-plot_regional_instruments_maxz}{}}}
\subsection{Method \code{plot_regional_instruments_maxz()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$plot_regional_instruments_maxz(
  instrument_region_zscores = self$instrument_region_zscores,
  instruments = self$instrument_raw,
  region = 1:min(10, nrow(instruments)),
  comparison = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-regional_ld_matrices"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-regional_ld_matrices}{}}}
\subsection{Method \code{regional_ld_matrices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$regional_ld_matrices(
  instrument_regions = self$instrument_regions,
  bfiles = self$bfiles,
  pops = self$pops,
  plink = self$plink
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-susie_finemap_regions"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-susie_finemap_regions}{}}}
\subsection{Method \code{susie_finemap_regions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$susie_finemap_regions(
  dat = self$instrument_regions,
  ld = self$ld_matrices
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-paintor_finemap_regions"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-paintor_finemap_regions}{}}}
\subsection{Method \code{paintor_finemap_regions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$paintor_finemap_regions(
  region = self$instrument_regions,
  ld = self$ld_matrices,
  PAINTOR = "PAINTOR",
  workdir = tempdir()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-MsCAVIAR_finemap_regions"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-MsCAVIAR_finemap_regions}{}}}
\subsection{Method \code{MsCAVIAR_finemap_regions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$MsCAVIAR_finemap_regions(
  region = self$instrument_regions,
  ld = self$ld_matrices,
  MsCAVIAR = "MsCAVIAR",
  workdir = tempdir()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-fema_regional_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-fema_regional_instruments}{}}}
\subsection{Method \code{fema_regional_instruments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$fema_regional_instruments(
  method = "fema",
  instrument_regions = self$instrument_regions,
  instrument_raw = self$instrument_raw,
  n = self$exposure_metadata$sample_size
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-plot_regional_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-plot_regional_instruments}{}}}
\subsection{Method \code{plot_regional_instruments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$plot_regional_instruments(
  region,
  instrument_regions = self$instrument_regions,
  meta_analysis_regions = self$instrument_fema_regions
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$get_metadata(
  exposure_ids = self$exposure_ids,
  outcome_ids = self$outcome_ids
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-estimate_instrument_heterogeneity_per_variant"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-estimate_instrument_heterogeneity_per_variant}{}}}
\subsection{Method \code{estimate_instrument_heterogeneity_per_variant()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$estimate_instrument_heterogeneity_per_variant(dat = self$harmonised_dat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-mrgxe"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-mrgxe}{}}}
\subsection{Method \code{mrgxe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$mrgxe(
  dat = self$harmonised_dat,
  variant_list = subset(self$instrument_heterogeneity_per_variant, Qfdr < 0.05)$SNP,
  nboot = 100
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-mrgxe_plot"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-mrgxe_plot}{}}}
\subsection{Method \code{mrgxe_plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$mrgxe_plot(mrgxe_res = self$mrgxe_res)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-mrgxe_plot_variant"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-mrgxe_plot_variant}{}}}
\subsection{Method \code{mrgxe_plot_variant()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$mrgxe_plot_variant(
  variant = self$mrgxe_res \%>\% dplyr::filter(p.adjust(a_pval, "fdr") < 0.05) \%>\% {
  
      .$SNP
 },
  dat = self$harmonised_dat
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-make_outcome_data"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-make_outcome_data}{}}}
\subsection{Method \code{make_outcome_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$make_outcome_data(exp = self$instrument_raw, p_exp = 0.05/nrow(exp))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-make_outcome_local"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-make_outcome_local}{}}}
\subsection{Method \code{make_outcome_local()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$make_outcome_local(
  exp = self$instrument_raw,
  out = self$instrument_outcome_regions,
  p_exp = 0.05/nreow(exp)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-harmonise"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-harmonise}{}}}
\subsection{Method \code{harmonise()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$harmonise(exp = self$instrument_raw, out = self$instrument_outcome)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-set_summary"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-set_summary}{}}}
\subsection{Method \code{set_summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$set_summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-pleiotropy"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-pleiotropy}{}}}
\subsection{Method \code{pleiotropy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$pleiotropy(harmonised_dat = self$harmonised_dat, mrres = self$mrres)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-plot_pleiotropy"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-plot_pleiotropy}{}}}
\subsection{Method \code{plot_pleiotropy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$plot_pleiotropy(dat = self$pleiotropy_outliers)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-plot_pleiotropy_heterogeneity"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-plot_pleiotropy_heterogeneity}{}}}
\subsection{Method \code{plot_pleiotropy_heterogeneity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$plot_pleiotropy_heterogeneity(
  dat = self$pleiotropy_Q_outliers,
  pthresh = 0.05
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-perform_basic_sem"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-perform_basic_sem}{}}}
\subsection{Method \code{perform_basic_sem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$perform_basic_sem(harmonised_dat = self$harmonised_dat_sem)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-runsem"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-runsem}{}}}
\subsection{Method \code{runsem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$runsem(model, data, modname)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-standardise_data"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-standardise_data}{}}}
\subsection{Method \code{standardise_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$standardise_data(
  dat = self$instrument_raw,
  standardise_unit = FALSE,
  standardise_scale = FALSE,
  scaling_method = "simple_mode"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
