% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organise_local_data.R
\name{CAMERA_local}
\alias{CAMERA_local}
\title{CAMERA_local class}
\description{
A simple wrapper function for importing data from local files for use with the CAMERA class.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CAMERA_local-new}{\code{CAMERA_local$new()}}
\item \href{#method-CAMERA_local-standardise}{\code{CAMERA_local$standardise()}}
\item \href{#method-CAMERA_local-read_file}{\code{CAMERA_local$read_file()}}
\item \href{#method-CAMERA_local-pool_tophits}{\code{CAMERA_local$pool_tophits()}}
\item \href{#method-CAMERA_local-organise_data}{\code{CAMERA_local$organise_data()}}
\item \href{#method-CAMERA_local-fixed_effects_meta_analysis_fast}{\code{CAMERA_local$fixed_effects_meta_analysis_fast()}}
\item \href{#method-CAMERA_local-organise}{\code{CAMERA_local$organise()}}
\item \href{#method-CAMERA_local-clone}{\code{CAMERA_local$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA_local-new"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA_local-new}{}}}
\subsection{Method \code{new()}}{
Create a new dataset and initialise an R interface
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA_local$new(
  metadata,
  ld_ref,
  plink_bin,
  mc.cores = 1,
  radius = 25000,
  pthresh = 5e-08,
  minmaf = 0.01
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{Data frame with information about the data. One row per dataset. See details for info on columns}

\item{\code{ld_ref}}{Data frame with two columns - pop = population (referencing the pop values in metadata), bfile = path to plink file for that reference}

\item{\code{plink_bin}}{Location of executable plink (ver.1.90 is recommended)}

\item{\code{radius}}{Genomic window size to extract SNPs}

\item{\code{pthresh}}{P-value threshold for instrument inclusion}

\item{\code{minmaf}}{Minimum allelel frequency per dataset}

\item{\code{clump_pop}}{Reference population for clumping}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA_local-standardise"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA_local-standardise}{}}}
\subsection{Method \code{standardise()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA_local$standardise(
  d,
  ea_col = "ea",
  oa_col = "oa",
  beta_col = "beta",
  eaf_col = "eaf",
  chr_col = "chr",
  pos_col = "pos",
  vid_col = "vid"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA_local-read_file"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA_local-read_file}{}}}
\subsection{Method \code{read_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA_local$read_file(m, minmaf = 0.01)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA_local-pool_tophits"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA_local-pool_tophits}{}}}
\subsection{Method \code{pool_tophits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA_local$pool_tophits(
  rawdat,
  tophits,
  metadata,
  radius = 250000,
  pthresh = 5e-08,
  mc.cores = 10
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA_local-organise_data"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA_local-organise_data}{}}}
\subsection{Method \code{organise_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA_local$organise_data(
  metadata = self$metadata,
  plink_bin = self$plink_bin,
  ld_ref = self$ld_ref,
  pthresh = self$pthresh,
  minmaf = self$minmaf,
  radius = self$radius,
  mc.cores = self$mc.cores
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA_local-fixed_effects_meta_analysis_fast"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA_local-fixed_effects_meta_analysis_fast}{}}}
\subsection{Method \code{fixed_effects_meta_analysis_fast()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA_local$fixed_effects_meta_analysis_fast(beta_mat, se_mat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA_local-organise"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA_local-organise}{}}}
\subsection{Method \code{organise()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA_local$organise()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CAMERA_local-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CAMERA_local-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CAMERA_local$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
