% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAncestrySummarySet.R
\name{MultiAncestrySummarySet}
\alias{MultiAncestrySummarySet}
\title{R6 class for MR.TRAM}
\description{
A simple wrapper function.
Using a summary set, identify set of instruments for the traits, and peform SEM MR to test the association across the population.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-import}{\code{MultiAncestrySummarySet$import()}}
\item \href{#method-new}{\code{MultiAncestrySummarySet$new()}}
\item \href{#method-extract_instruments}{\code{MultiAncestrySummarySet$extract_instruments()}}
\item \href{#method-check_phenotypes}{\code{MultiAncestrySummarySet$check_phenotypes()}}
\item \href{#method-standardise_units}{\code{MultiAncestrySummarySet$standardise_units()}}
\item \href{#method-extract_instrument_regions}{\code{MultiAncestrySummarySet$extract_instrument_regions()}}
\item \href{#method-estimate_instrument_specificity}{\code{MultiAncestrySummarySet$estimate_instrument_specificity()}}
\item \href{#method-scan_regional_instruments}{\code{MultiAncestrySummarySet$scan_regional_instruments()}}
\item \href{#method-plot_regional_instruments}{\code{MultiAncestrySummarySet$plot_regional_instruments()}}
\item \href{#method-regional_ld_matrices}{\code{MultiAncestrySummarySet$regional_ld_matrices()}}
\item \href{#method-susie_finemap_regions}{\code{MultiAncestrySummarySet$susie_finemap_regions()}}
\item \href{#method-paintor_finemap_regions}{\code{MultiAncestrySummarySet$paintor_finemap_regions()}}
\item \href{#method-run_MsCAVIAR}{\code{MultiAncestrySummarySet$run_MsCAVIAR()}}
\item \href{#method-plot_finemapping}{\code{MultiAncestrySummarySet$plot_finemapping()}}
\item \href{#method-make_outcome_data}{\code{MultiAncestrySummarySet$make_outcome_data()}}
\item \href{#method-harmonised_dat}{\code{MultiAncestrySummarySet$harmonised_dat()}}
\item \href{#method-perform_basic_sem}{\code{MultiAncestrySummarySet$perform_basic_sem()}}
\item \href{#method-clone}{\code{MultiAncestrySummarySet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import"></a>}}
\if{latex}{\out{\hypertarget{method-import}{}}}
\subsection{Method \code{import()}}{
Migrate the results from a previous MultiAncestrySummarySet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$import(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new dataset and initialise an R interface
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$new(
  exposure_ids = NULL,
  outcome_ids = NULL,
  pops = NULL,
  bfiles = NULL,
  plink = NULL,
  radius = NULL,
  clump_pop = NULL,
  x = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exposure_ids}}{IDs for the exposures from each population.}

\item{\code{outcome_ids}}{IDs for the outcomes from each population.}

\item{\code{pops}}{Ancestry information; AFR, AMR, EUR, EAS, SAS}

\item{\code{bfiles}}{Location of LD reference file (Download from: http://fileserve.mrcieu.ac.uk/ld/1kg.v3.tgz)}

\item{\code{plink}}{Location of executable plink (ver.1.90)}

\item{\code{radius}}{Set a range of the region to look for}

\item{\code{clump_pop}}{insert}

\item{\code{x}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-extract_instruments}{}}}
\subsection{Method \code{extract_instruments()}}{
Identifies the instruments for the exposure
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$extract_instruments(
  exposure_ids = self$exposure_ids,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exposure_ids}}{ID for the exposure. Default is x$exposure_ids.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_phenotypes"></a>}}
\if{latex}{\out{\hypertarget{method-check_phenotypes}{}}}
\subsection{Method \code{check_phenotypes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$check_phenotypes(
  ids = self$exposure_ids,
  after_standardised = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standardise_units"></a>}}
\if{latex}{\out{\hypertarget{method-standardise_units}{}}}
\subsection{Method \code{standardise_units()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$standardise_units(exp = NULL, out = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract_instrument_regions"></a>}}
\if{latex}{\out{\hypertarget{method-extract_instrument_regions}{}}}
\subsection{Method \code{extract_instrument_regions()}}{
Extract a region around each instrument obtained from \code{x$extract_instruments()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$extract_instrument_regions(
  radius = self$radius,
  instrument_raw = self$instrument_raw,
  exposure_ids = self$exposure_ids
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{radius}}{Set a range of the region to search}

\item{\code{instrument_raw}}{A set of instruments obtained from \code{x$extract_instruments()}}

\item{\code{exposure_ids}}{ID for the exposure. Default is x$exposure_ids.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimate_instrument_specificity"></a>}}
\if{latex}{\out{\hypertarget{method-estimate_instrument_specificity}{}}}
\subsection{Method \code{estimate_instrument_specificity()}}{
Estimate what fraction of the instuments is expected to be replicated under the hypothesis that the effect estimates are the same
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$estimate_instrument_specificity(
  instrument,
  alpha = "bonferroni"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instrument}}{A set of instruments obtained from \code{x$extract_instruments()} or \code{scan_regional_instruments()}}

\item{\code{alpha}}{Significance level. Default is 0.05/number of SNPs (Bonferroni)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scan_regional_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-scan_regional_instruments}{}}}
\subsection{Method \code{scan_regional_instruments()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$scan_regional_instruments(
  instrument_raw = self$instrument_raw,
  instrument_regions = self$instrument_regions
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instrument_raw}}{insert}

\item{\code{instrument_regions}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_regional_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-plot_regional_instruments}{}}}
\subsection{Method \code{plot_regional_instruments()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$plot_regional_instruments(
  region = 1:min(10, nrow(instruments)),
  instrument_region_zscores = self$instrument_region_zscores,
  instruments = self$instrument_raw
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{insert}

\item{\code{instrument_region_zscores}}{insert}

\item{\code{instruments}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-regional_ld_matrices"></a>}}
\if{latex}{\out{\hypertarget{method-regional_ld_matrices}{}}}
\subsection{Method \code{regional_ld_matrices()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$regional_ld_matrices(
  instrument_regions = self$instrument_regions,
  bfiles = self$bfiles,
  pops = self$pops,
  plink = self$plink
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instrument_regions}}{insert}

\item{\code{bfiles}}{insert}

\item{\code{pops}}{insert}

\item{\code{plink}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-susie_finemap_regions"></a>}}
\if{latex}{\out{\hypertarget{method-susie_finemap_regions}{}}}
\subsection{Method \code{susie_finemap_regions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$susie_finemap_regions(
  dat = self$instrument_regions,
  ld = self$ld_matrices
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-paintor_finemap_regions"></a>}}
\if{latex}{\out{\hypertarget{method-paintor_finemap_regions}{}}}
\subsection{Method \code{paintor_finemap_regions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$paintor_finemap_regions(
  region = self$instrument_regions,
  ld = self$ld_matrices,
  PAINTOR = "PAINTOR",
  workdir = tempdir()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{workdir}}{Working directory. Default=tempdir()}

\item{\code{regiondata}}{Output from extract_regional_data}

\item{\code{mscaviar}}{Path to mscaviar executable. Default="mscaviar"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Results table with posterior inclusion probabilities
Analyse regional data with MsCAVIAR
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_MsCAVIAR"></a>}}
\if{latex}{\out{\hypertarget{method-run_MsCAVIAR}{}}}
\subsection{Method \code{run_MsCAVIAR()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$run_MsCAVIAR(
  region = self$instrument_regions,
  ld = self$ld_matrices,
  MsCAVIAR = "/work/yc16575/MsCAVIAR/MsCAVIAR",
  workdir = tempdir()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MsCAVIAR}}{Path to MsCAVIAR executable. Default="MsCAVIAR"}

\item{\code{workdir}}{Working directory. Default=tempdir()}

\item{\code{regiondata}}{Output from extract_regional_data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Results table with posterior inclusion probabilities
Plot paintor results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_finemapping"></a>}}
\if{latex}{\out{\hypertarget{method-plot_finemapping}{}}}
\subsection{Method \code{plot_finemapping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$plot_finemapping(
  instruments = self$instrument_paintor,
  probability_scores = self$paintor_results,
  region = 1:min(10, nrow(instruments))
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{res}}{Output from run_paintor}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_outcome_data"></a>}}
\if{latex}{\out{\hypertarget{method-make_outcome_data}{}}}
\subsection{Method \code{make_outcome_data()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$make_outcome_data(exp = self$instrument_raw, p_exp = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exp}}{insert}

\item{\code{p_exp}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-harmonised_dat"></a>}}
\if{latex}{\out{\hypertarget{method-harmonised_dat}{}}}
\subsection{Method \code{harmonised_dat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$harmonised_dat(
  exp = self$instrument_raw,
  out = self$instrument_outcome
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-perform_basic_sem"></a>}}
\if{latex}{\out{\hypertarget{method-perform_basic_sem}{}}}
\subsection{Method \code{perform_basic_sem()}}{
insert
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$perform_basic_sem(
  harmonised_dat = self$harmonised_dat_sem
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{harmonised_dat}}{insert}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiAncestrySummarySet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
